Make the search form to do something
I am implementing a function for handling submitted search form on
the home page and actually try to search something.

We need to handle it on some URL and then redirect to the expected
`/packages/s/foo` but in the end, we should have a JS version for
doing it more nicely.

I am mocking both package model class and a way of obtaining the
list of packages from the database because I don't want to deal with
that right now.